Migrated scroll and click events to scroll-load-more.helpers.js file from MainfeedScreen component.
