+++ src/screens/MainFeedScreen/MainFeedScreen.js
@@ -13,6 +13,7 @@ import { Link } from "react-router-dom";
 import IsLogged from "../../components/IsLogged/IsLogged";
 import LoginModal from "../../components/LoginModal/LoginModal";
 import ContentLoader from "../../components/ContentLoader/ContentLoader";
+import "../../components/_helpers/scroll-load-more.helpers";
 
 export const MainFeedContext = React.createContext();
 
@@ -22,28 +23,6 @@ const MainFeedScreen = () => {
   const user = useSelector((state) => state.user);
   const data = useSelector((state) => state.data);
   const { loading, articles } = data;
-
-  window.addEventListener("scroll", () => {
-    let aside = document.getElementsByClassName("home--box-8")[0];
-    if (aside) {
-      if (window.scrollY > 400) {
-        aside.classList.add("home--fixed");
-      } else {
-        aside.classList.remove("home--fixed");
-      }
-    }
-  });
-
-  window.addEventListener("click", (e) => {
-    let backdrop = document.getElementsByClassName("actionmenu")[0];
-    let drawer = document.getElementsByClassName("actionmenu")[0];
-    if (e.target === backdrop) {
-      drawer.classList.remove("actionmenu--open");
-      drawer.classList.add("actionmenu--close");
-      backdrop.classList.remove("actionmenu--open");
-      backdrop.classList.add("actionmenu--close");
-    }
-  });
 
   useEffect(() => {
     dispatch(articlesActions.load({ category: "" }));
